스코어 UI수정 작업 중

+++ app/src/main/java/com/eye3/golfpay/fmb_tab/adapter/NearestLongestAdapter.java
@@ -71,7 +71,7 @@ public class NearestLongestAdapter extends RecyclerView.Adapter<NearestLongestAd
         }
     }
 
-    public void setScore(int score) {
+    public void setScore(float score) {
 
         if (score == -1)
             return;
